+++ src/main/java/co/empathy/IMDBproject/Model/Query/Filters.java
@@ -1,4 +1,4 @@
-package co.empathy.IMDBproject.Model;
+package co.empathy.IMDBproject.Model.Query;
 
 import lombok.Builder;
 import lombok.Value;
