+++ src/main/java/co/empathy/IMDBproject/Model/Response.java
@@ -8,6 +8,7 @@ import lombok.NoArgsConstructor;
 import java.util.ArrayList;
 import java.util.List;
 
+
 @Data
 @NoArgsConstructor
 public class Response {
