+++ src/main/java/co/empathy/IMDBproject/Service/ElasticServiceImpl.java
@@ -2,12 +2,12 @@ package co.empathy.IMDBproject.Service;
 
 import co.empathy.IMDBproject.Help.IMDbReader;
 import co.empathy.IMDBproject.Help.IMDbData;
-import co.empathy.IMDBproject.Model.Facets.Facets;
-import co.empathy.IMDBproject.Model.Filters;
+import co.empathy.IMDBproject.Model.Query.Filters;
 import co.empathy.IMDBproject.Model.Movie.Movie;
 
 import co.empathy.IMDBproject.Model.Response;
-import co.empathy.IMDBproject.Model.Sort;
+import co.empathy.IMDBproject.Model.Query.Sort;
+import org.springframework.core.io.ClassPathResource;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.*;
@@ -51,13 +51,15 @@ public class ElasticServiceImpl implements ElasticService {
     public Boolean indexIMDBData(MultipartFile basicsFile, MultipartFile ratingFile, MultipartFile akasFile,MultipartFile crewFile,MultipartFile principalsFile) throws IOException {
 
         try {
+            //checks that there are not empty files
             if (basicsFile.isEmpty()||ratingFile.isEmpty()||akasFile.isEmpty()||crewFile.isEmpty()||principalsFile.isEmpty())
                 throw new IOException();
 
             imdb = new IMDbReader(basicsFile, ratingFile, akasFile, crewFile, principalsFile);
+            //starts reading the first lines
             imdb.initializeLines();
             //create imdb index and set the mapping properties
-            createIndex(imdbIndex, data.jsonMapping());
+            createIndex(imdbIndex, jsonMapping());
 
 
             List<Movie> movieList = new ArrayList<>();
@@ -76,7 +78,6 @@ public class ElasticServiceImpl implements ElasticService {
                     //index a "small" movie's list
                     elasticEngine.indexMultipleDocs(imdbIndex, movieList);
 
-
                     //prepare the next list
                     countMovies = 0;
                     movieList.clear();
@@ -85,7 +86,7 @@ public class ElasticServiceImpl implements ElasticService {
             //index the last list if is not empty
 
             elasticEngine.indexMultipleDocs(imdbIndex, movieList);
-            System.out.println("Indexed");
+
         }
         catch (IOException e){
             throw e;
@@ -103,6 +104,16 @@ public class ElasticServiceImpl implements ElasticService {
     public Response getSearchQuery(String searchText) throws IOException {
         return elasticEngine.getSearchQuery(searchText);
 
+    }
+
+    /**
+     *
+     * @return an input stream with the .json from the resources folder
+     * @throws IOException
+     */
+    public InputStream jsonMapping() throws IOException {
+        InputStream mappingInputStream = new ClassPathResource("static/mapping.json").getInputStream();
+        return mappingInputStream;
     }
