+++ src/main/java/co/empathy/IMDBproject/Service/QueryProvider.java
@@ -1,16 +1,11 @@
 package co.empathy.IMDBproject.Service;
 
 import co.elastic.clients.elasticsearch._types.FieldValue;
-import co.elastic.clients.elasticsearch._types.ScoreSort;
-import co.elastic.clients.elasticsearch._types.SortOptions;
-import co.elastic.clients.elasticsearch._types.SortOrder;
 import co.elastic.clients.elasticsearch._types.query_dsl.*;
 import co.elastic.clients.json.JsonData;
-import co.empathy.IMDBproject.Model.Filters;
+import co.empathy.IMDBproject.Model.Query.Filters;
 
 
-
-import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
@@ -20,6 +15,12 @@ import static java.util.Objects.nonNull;
 public class QueryProvider {
 
     //queryBuilder
+
+    /**
+     *
+     * @param filters you want to apply
+     * @return a query to send in the request with all the filters
+     */
     public Query getFilterQuery(Filters filters){
 
         List<Query> queries=new ArrayList<>();
@@ -40,12 +41,11 @@ public class QueryProvider {
         }
         if(nonNull(filters.getMinYear()) && nonNull(filters.getMaxYear())){
 
-            //using always start year
+
             queries.add(rangeQuery(filters.getMinYear(),filters.getMaxYear(),"startYear"));
         }
 
         if(nonNull(filters.getMaxScore()) && nonNull(filters.getMinScore())){
-            //order result by num of votes?
 
             queries.add(rangeQuery(filters.getMinScore(),filters.getMaxScore(),"averageRating"));
         }
@@ -57,14 +57,27 @@ public class QueryProvider {
 
 
     }
+    /**
+     *
+     * @param value, the exact term you wish to find
+     * @param field you wish to search
+     * @return the query that will be sent in the request
+     */
 
     public Query termQuery(String value, String field){
-        Query query = TermQuery.of(m -> m
+        return  TermQuery.of(m -> m
                 .field(field)
                 .value(value)
         )._toQuery();
-        return query;
+
     }
+
+    /**
+     *
+     * @param values, an array with the multiple values
+     * @param field you wish to search
+     * @return the query that will be sent in the request
+     */
 
     public Query termsQuery(String[] values, String field){
 
@@ -72,13 +85,20 @@ public class QueryProvider {
         TermsQueryField terms = new TermsQueryField.Builder()
                 .value(list.stream().map(FieldValue::of).toList())
                 .build();
-        Query query = TermsQuery.of(m -> m
+        return TermsQuery.of(m -> m
                 .field(field)
                 .terms(terms)
         )._toQuery();
-        return query;
+
     }
 
+    /**
+     *
+     * @param minValue, the min value of the term (integer)
+     * @param maxValue, the max value of the term (integer)
+     * @param field you wish to search
+     * @return the query that will be sent in the request
+     */
     public Query rangeQuery(int minValue, int maxValue, String field){
         return RangeQuery.of(r -> r
                 .field(field)
@@ -87,6 +107,13 @@ public class QueryProvider {
         )._toQuery();
 
     }
+    /**
+     *
+     * @param minValue, the min value of the term (double)
+     * @param maxValue, the max value of the term (double)
+     * @param field you wish to search
+     * @return the query that will be sent in the request
+     */
     public Query rangeQuery(Double minValue, Double maxValue, String field){
         return RangeQuery.of(r -> r
                 .field(field)
@@ -95,6 +122,13 @@ public class QueryProvider {
         )._toQuery();
 
     }
+
+    /**
+     *
+     * @param value, the query string
+     * @param fields, the fields to be queried
+     * @return the query that will be sent in the request
+     */
     public Query multiMatchquery(String value, List <String> fields)  {
         return MultiMatchQuery.of(r -> r
                 .fields(fields)
